Switch Snapshot flows to fetch HOI by clientID
This makes the actual switch in the Add/Remove person flows to
fetch HOI by client IDs. It updates the expectations of most
snapshot feature tests to stub the new endpoint. HOWEVER, it does
not fix the expectations of the HOI feature test itself, as there
are more significant changes that need to happen there.