Update HOI feature specs
This updates the screening and snapshot feature specs to handle
Snapshot fetching HOI from Ferb by client ids. This is an
unfortunately large change, as we can no longer depend on the
feature flag :hoi_from_intake_api to predict what our data will
look like.

This also includes a small patch to the clientIds selector, which
was not falling back to legacy_descriptor properly when legacy_id
was set explicitly to null.