Fetch Snapshot person data without POSTing participant
This updates the CREATE_SNAPSHOT_PERSON action to receive a single
ID parameter for the client being added. That ID is used to fetch
the person summary data for the client, without adding them to a
persistent snapshot in Postgres.

Note that fetching the client data is still necessary. Although
many times the data already exists from a search the user ran, they
can also add clients through the Relationships Card (and maybe
other future places).