Remove the API call when deleting snapshot people
Snapshot participants are no longer persisted to Postgres, and so
deleting them does not require any special API call. We simply
fire off the event that removing them succeeded (so they are no
longer in the client list) and then fetch new rels and hois.
Because there is no API call, we no longer need to catch those
errors.