Convert from dora format in CreateSnapshotPersonSaga
This should not live in the Saga, but we have no better place for
it right now. Ideally, reducers should serve the purpose of
normalizing data on its way into the Redux store. However,
reducers should also be pure, and our data mappers require full
access to the state for their LOV values. So, for now we push this
to the Saga, which can freely select from the state and cause side
effects. This has the benefit of reducing how far into the frontend
we allow Dora results to traverse.