Display Intersex gender in the frontend
This adds Intersex to our gender enum, to allow it to be displayed
when we are interacting with clients containing that info in the
backend. Note that 'gender' here is the future 'sex_at_birth'
field. Also note that we are not allowing Intersex to be pushed to
the backend, so undefined behavior may occur when editing Intersex
client information.