Track pending snapshot people additions in Redux
Whenever we attempt to add a snapshot person, we add that person
to the redux store as a pendingParticipant. This will allow us to
reactively update the UI based on that state, so we can do things
like remove the Attach link for that person and prevent duplicate
action calls.