Float the HOI table index to the left padding gutter
This adds a new table style 'ordered-table' (named after ol -
ordered list), which styles its first column of indices to float
in the left hand gutter of its parent (in this case, an HOI card).
This primarily saves a little space for this column, while also
making the numbers stand out just a bit more.

Manually tested with numbers up to 99 in Chrome, Firefox, and IE.