Make cards position relative
We use cards to represent fairly contained pieces of the UI, and so
it makes sense for them to be positioned relatively, such that any
absolutely positioned children compute their location relative to
the card. This is a little less hacky than making the HOI card
alone relative, which hit against a lint rule.