Abstract existing optimistic auth check
When searching for clients to add to a snapshot or screening, we
currently do a preliminary authorization check on the client to see
if the user has any chance of being able to add the client. We will
be expanding this check to cover more cases, and so the first step
is to abstract it into its own, testable file.