Pass client county down through search results
This adds the client county code to each client object returned in
Dora person_summary search results. We are not rendering this info
at the moment, but it will soon become useful in client-side auth
checks. (Which, to always reiterate, are in addition to server-side
checks, not in place of.)