Check counties when adding clients
When attempting to add a client to a snapshot or screening, we do
a preliminary authorization check to see if that action is even
possibly going to be allowed. This commit expands that logic to
check whether the users are in the same county or not, unless we
detect that the user has any of several override authority
privileges present.