Refactor rendering of search result names
Rather than pass individual, possibly emphasized name fields (for
first_name, middle_name, last_name, and name_suffix) down to the
PersonSuggestion component, we instead pass a fully formatted full
name. The tradeoff for this is that more logic is handled in the
peopleSearchSelectors, including detecting whether a suffix needs
a comma or whether to display the suffix at all. The tradeoff is
worthwhile, though, as it allows us to avoid a massive enum of
emphasized suffixes.