Format highlighted name suffixes
If Dora gives us search results with a highlighted name suffix, we
format that name suffix (as though it did not have the emphasis)
and then wrap the result in new highlight emphasis tags. This also
has the effect of only displaying what we consider valid suffixes,
regardless of the source (Ferb/Dora).