Move code around in ScreeningPage
This commit tightens up a few lines of code in ScreeningPage to
bring it in line with Code Climate analysis. This is obviously not
a sustainable way of fixing this file. A file this large should be
broken down into a smaller components. One easy way to do this
would be to create card components for each of the form/show
combos. Such a change would move and change many files, though, and
so it should wait until it can be coordinated across the team.