Default CrossReport agency fields to first value
When editing the Cross Report fields for a screening, once the user
clicks to enable a particular type of agency, we default to the
first valid option of that agency type. This often removes a click
for the user, if that field is the one they wanted.

Since the empty option is invalid anyways, and we are defaulting to
a value, I went ahead and removed some of the validation tests for
those errors in feature tests. The errors should no longer happen,
although I am leaving the validation itself in case somebody
decides to 'revert' this behavior without actually reverting the
whole commit.