Remove pending basepath test for logout
This removes the pending basepath test for logout, which was flakey
and passing or failing depending on the particular order that tests
were run. This is fragile when we are parallelizing our test suite,
and we think it is related to the global Rails config settings in
this file. So, we remove this for now and hope to test basepath
functionality in other ways.