HOT-1880 Filter addresses without legacy descriptors at the reducer
Address legacy_descriptor is empty when creating the participant,
but comes populated when loading the screening.
Quite a head scratcher... common denominator seems to be that legacy_id
is always present, so the reducer filters by that.

I'd change the API, but this will be touched again when ferb starts
saving participants and it will probably be addressed then.

I've created the following story to work on it later https://osi-cwds.atlassian.net/browse/HOT-1948