Set max-size of avatar picture
This sets the max-size of the avatar picture on the person info
card. This covers the screensize band from 0 to 991px, or more
accurately from 768px to 991px, since react-wood-duck already set
the same max size on smaller screens.

This is necessary to patch here, since we are using a _slightly_
different card component to react-wood-duck, but some of the shared
styles do still apply here.