Reduce Ferb allegations to expected format
When we switched from Intake API to Ferb for our screening GETs, we
had to add some mappings for the slightly changed API responses.
We missed one piece, though, which is that allegations are actually
reduced in multiple places. This was causing bad data to enter our
store, and eventually caused selectors to crash.