Format SSB info for Show fields, not for Form fields
This represents the complete cycle of using system codes to store
SSB values. The form uses the raw values to select the options and
saves those codes to the Redux store. The Show component translates
those (actually in the Selector) into proper strings for display.

This commit is a major milestone in this SSB branch.