Update the SSB show when you save the person
Now that the state for persisted and live/form data for SSB are
split, we add an action that can "save" the data to the persisted
state. This uses a saga that will call to update the actual data in
Postgres, but for now, that API call is not available. So we do all
we can and copy the data to the store immediately.