Display SSB surrendered by when perps are set
When a participant is updated, we check to see if they have the
perpetrator role. If they do, and the report type is SSB, then they
are presumed to be the one who surrendered the baby. This was
tested manually through some unstaged changes, as the report_type
isn't actually able to be set right now. Unit tests pass as well
of course.