Post participant data for Baby Doe and Unknown Doe
This doesn't quite work yet. We add the participants, but most/all
of the fields we attempt to set are not actually created. They get
filtered out by the Participant Repository layer in Rails. Before
fixing that, I want to refactor so that the few fields we do set
can be passed straight through that layer (and then all params can
pass through as well).