Add Ferb endpoint for posting participants
This adds the Ferb endpoint for posting participants, which is not
in use yet but will be used by SSB to post premade Baby Does. The
endpoint is defined as a screenings endpoint, because that is where
the proper endpoint will exist. For now, it uses the compatible
(but temporary) /participants endpoint.

This also removes unnecessary stripping of params when creating
participants, since we already have permitted params check.