Use Ferb when creating new (non-legacy) participants
This updates the Participant repository and feature tests to use
the Ferb participants endpoint for creating new participants with
fields such as first name and last name specified. This endpoint
is not yet in its final location, nor can it handle existing
clients when legacy_descriptors are specified. But, it works for
now, for SSB purposes, and this is not moving further away from
where we want to be.