Handle edge cases in generateBabyDoeSaga
This handles two edge cases in the saga to generate data for SSB
reports. First, it catches API errors when creating participants
and handles them as real createPersonFailures, so if they fail the
user should get notified. It also handles a potential (but in
practice, nonexistant) race condition with the setAllegations
action, to ensure it is fully dispatched (and processed by
reducers) before moving on to save the screening.