Add feature test for generating Baby Doe
Adding an end-to-end test for generating SSB data also turned up a
handful of bugs, such as treating surrenderedBy as a participant ID
when it is intended to be a string of the person's name ("Unknown
Doe"). I also added some error handling to the saga and removed
some dead code.