Migrate to Ferb endpoint for screening submit
This moves us to the Ferb endpoint for screening submit. Since the
endpoint is compatible with the existing Intake endpoint as well
as the Ferb endpoint we use to fetch screening info, this is
practically a URL change. There is a small amount of processing of
the response that had to change, to deal with an extra unexpected
field (incident_address).