Save screeningInfo within generateBabyDoeSaga
This incorporates the saving of the screening info (including the
report type) into the generateBabyDoeSaga. This resolves a race
condition, where creating the new participants sometimes happened
faster than this initial screening update. When that occurred, the
screening update response would contain an outdated list of
participants.

This PR duplicates some logic between the generateBabyDoeSaga and
the saveScreeningCardSaga. This is primarily because using yield*
to call sagas from each other results in painful unit tests. I am
asynchronously considering ways to improve this, but it is outside
the scope of this story.