Update allegationsForm on any screening update
Prior to this, we only updated the allegations form with new
allegations when fetching a screening. This is because we assumed
that if the allegations changed, it must have been due to the user
submitting the form. This was never true, but particularly now when
allegations can be saved by triggering a Baby Doe generation.

Practically speaking, this prevents a bug where submitting the
allegations form after setting report type to SSB causes a 500
error, because it attempts to create the same allegation twice.