Do not store participants on screening in Redux
This removes the participants property from the screening in Redux,
to avoid confusion. This participants list should not have been
used, as the correct, up-to-date list of participants was stored
separately. To make sure we don't accidentally use this bad
participants list, we avoid storing it in the first place.