Abort updating the screening if participants are undefined
This primarily guards against the case where we select a screening
without merging in from the participants List in the Redux state.
While normally guarding like this would be unnecessary, the
potential for data loss is so drastic that we need to detect and
alert when this happens. The alternative is to quietly delete all
participants from the screening, which would have huge negative
consequences for the user.