HOT-1529 Added an extra field on the decision form to hold the screening contact reference.
A Screening Contact is a different entity in legacy (it is NOT A
REFERRAL).

It is an additional piece of information that gets aggregated to an
existing Referral or Case.

To make the connection and have the info added to the contact to the
right Referral or Case, the id must be provided to be used as a parent
for the contact.