Remove safetyAlert reducer and just update hardcoded list
This is a workaround pending additional Ferb functionality. Each
systemCode row in the database contains a legacy system id, a new
intake_code, and a intake_display value with a human readable
display. Ferb endpoints expect us to send the intake_code, but this
code is not provided in the LOV service. So, the only option we
have at the moment is to hardcode it.