Preserve touched status on person update
The intended behavior of our forms is that validation messages
will be displayed after you save and re-edit a card. When saving,
all fields are marked as touched. This patches a bug where the
touch status was wiped on successful person update. Through a fluke
in our feature test timing, our tests were still passing. (The
proper long term fix for this is to add a Loading state to our
cards, so that we do not have race conditions with API calls.)