Click select fields to trigger focus, and other validation fixes
For some reason, capybara is able to change a select field value
without trigger that field to focus. If it never focuses, it never
blurs, and so our validation messages are not triggered as
expected.

Included in this commit are a couple other expectation or bug fixes
where the validations were not being displayed properly.