Add ScrollToTop component for managing focus on route change
ScrollToTop renders a focusable div and sits at the very top of our
App. Whenever the Route changes, it scrolls to top and focuses on
the div, thus resetting the focus to the top of the page.

The div itself is kept outside of the keyboard navigation flow with
tabIndex of -1. These are in line with current recommendations from
the React-Router community.