Turn on code duplication check for JS
This enables code duplication checks for JavaScript, at a level
that will not detect any *current* issues. Our highest issue has a
mass of 169. Turning this on will prevent worse duplication from
entering our codebase, and we can aim to lower the threshold over
time.

Ruby setting is 25, same as the default. JS default would be 45.