Reduce addresses on participants
This transforms addresses on participants to a unified format,
which I am currently calling 'plain addresses'. This is just the
standard set of fields that we use, not wrapped in any touch/value
construct.

This commit does not handle actually using these plain objects just
yet. Displaying addresses and saving cards is likely broken.