Update address feature tests
This updates a couple of our feature tests to expect the proper
fields for addresses, including that a legacy_descriptor will be
present, and that we will not post/put a legacy_id (always use the
descriptor). It also makes small changes to our helpers to ensure
that an id of null is always present explicitly in a new address.