Feature test Narrative Card saving state
There are later plans to refactor how the card is setting the state
itself. The goal of this was just to prove that we can feature test
this behavior. In order to do so, we need to add a small sleep to
the API stubs, because otherwise they finish too fast for Capybara
to see the spinner. We add a longer wait time on the next expect
statement to compensate, just in case.