Add a select field to search for county by name
This adds a select dropdown for counties to the search form when
address search is enabled. To keep the select field as
componentized as possible (and therefore avoid threading 'counties'
through multiple layers of components), I use a new injector
pattern which is a standalone container that 'loads' the counties
prop into its children. In future refactoring, this can be used to
make system code handling much easier throughout the application.